ae-bridge: Remove ASDP workaround.
ASDP = Activity Stack Double-Pop, the name we gave a particular bug
where the activity stack appears to be popped twice when the emulator
quits.

In SDL2 they identified the bug in issue 1293:
http://bugzilla.libsdl.org/show_bug.cgi?id=1293

and fixed it (in terms of SDL life) in changeset 6186 (e565ac981de6)
http://hg.libsdl.org/SDL/rev/e565ac981de6

As of this commit, emulator shutdown appears to be successful, though
when you relaunch the emulator, the whole activity crashes from a
segfault and the stack gets popped twice.  On some devices you can then
relaunch the emulator successfully.  And the cycle repeats.

One benefit as of this commit is that the logcat isn't lost when you
quit, so if you send a crash report it will actually contain data from
when the emulator was running.